requirements.txt
request
beautfulsoup4
lxml
